backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "static-api"
version = "0.4.0"
description = "A real-time web chat service using FastAPI with WebSockets."
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.12",
    "pydantic[email]>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
    "websockets>=16.0",
    "pytest-asyncio>=1.3.0",
    "fakeredis>=2.35.1",
    "testcontainers>=4.14.2",
    "httpx-ws>=0.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24.0",
    "anyio>=4.0.0",
    "httpx>=0.28.1",
    "testcontainers[postgres]>=4.8.0",
    "websockets>=16.0",
    "fakeredis[aioredis]>=2.26.0",
    "httpx-ws>=0.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "alembic*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
